+++ model_delivery_contract.go
@@ -82,7 +82,7 @@ type DeliveryContract struct {
 	// Current total long position size
 	PositionSize int64 `json:"position_size,omitempty"`
 	// Configuration's last changed time
-	ConfigChangeTime float32 `json:"config_change_time,omitempty"`
+	ConfigChangeTime float64 `json:"config_change_time,omitempty"`
 	// Contract is delisting
 	InDelisting bool `json:"in_delisting,omitempty"`
 	// Maximum number of open orders
